+++ R/run_indiv_efm_set.R
@@ -1,4 +1,14 @@
-#' Title Running individual EFM SNP sets in parallel
+# -------------------------------------------------------------------------------------------------
+# Copyright (c) 2024, DHS.
+#
+# This file is part of MixDeR and is licensed under the BSD license: see LICENSE.
+#
+# This software was prepared for the Department of Homeland Security (DHS) by the Battelle National
+# Biodefense Institute, LLC (BNBI) as part of contract HSHQDC-15-C-00064 to manage and operate the
+# National Biodefense Analysis and Countermeasures Center (NBACC), a Federally Funded Research and
+# Development Center.
+# -------------------------------------------------------------------------------------------------
+#' Running individual EFM SNP sets in parallel
 #'
 #' @param i Set number
 #' @param ids dataframe of replicate data